refactor(v010): carve voicegw project (detail) into projects.py
Sixth per-command carve-out in section 2.

Appends 'project' (the detail command) to voicegateway/cli/projects.py
alongside the listing command landed in the previous commit. Module
docstring updated to reflect the now-complete project surface.
Imports added: asyncio (for asyncio.run) and rich.panel.Panel (for
the project-detail card). The unused 'from rich.panel import Panel'
in _legacy.py is dropped in the same commit.

Implementation note: the PostToolUse formatter stripped the new
imports between my two Edit calls (the imports were unused at the
moment ruff saw the file, between Edit 1 and Edit 2). A third Edit
restored them. Documented in the journal so future carve-outs that
add both imports and consumer bodies sequence the body edit first.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package.
